oh-tab-ymg0: Extract E2E request helper
